Added Fluent Validation for employee and order patch methods

+++ TestAPIproject/Application/Services/EmployeeService.cs
@@ -26,19 +26,26 @@ namespace TestAPIproject.Application.Service
             var employees = await _repo.GetAllAsync(pagination.PageNumber, pagination.PageSize);
             return employees.Select(e => new EmployeeDto
             {
-                Id = e.EmployeeId,
-                Name = e.Name
+                EmployeeId = e.EmployeeId,
+                Name = e.Name,
+                Department = e.Department
             });
         }
 
-        public async Task<Employee?> GetByIdAsync(int id)
+        public async Task<EmployeeDto?> GetByIdAsync(int id)
         {
-            return await _repo.GetByIdAsync(id);
+
+            var employee = await _repo.GetByIdAsync(id);
+
+            if (employee == null)
+                return null;
+
+            return _mapper.Map<EmployeeDto>(employee);
         }
 
 
 
-        public async Task<Employee> AddAsync(EmployeeCreateDto model)
+        public async Task<EmployeeDto> AddAsync(EmployeeCreateDto model)
         {
 
             var employee = _mapper.Map<Employee>(model);
@@ -50,8 +57,8 @@ namespace TestAPIproject.Application.Service
             };*/
 
             await _repo.AddAsync(employee);
-            _logger.LogInformation(employee.EmployeeId, "added sccessfully");
-            return employee;
+            _logger.LogInformation("Employee with Id {EmployeeId} added successfully", employee.EmployeeId);
+            return _mapper.Map<EmployeeDto>(employee);
         }
 
         public async Task UpdateAsync(EmployeeEditDto model)
@@ -59,19 +66,52 @@ namespace TestAPIproject.Application.Service
             var employee = await _repo.GetByIdAsync(model.Id);
 
             if (employee == null)
-                throw new Exception("Employee not found");
+                throw new KeyNotFoundException("Employee not found");
 
             employee.Name = model.Name;
             employee.UpdateSalary(model.Salary);
+            employee.Department = model.Department;
+            employee.Role= model.Role;
+            employee.manager_id = model.Manager_Id;
+            await _repo.SaveAsync();
+            _logger.LogInformation("Employee with Id {EmployeeId} updated", model.Id);
+        }
 
 
+        public async Task PatchEmployeeAsync(int id, PatchEmployeeDto dto)
+        {
+            var employee = await _repo.GetByIdAsync(id);
+
+            if (employee == null)
+                throw new Exception("Employee not found");
+
+            if (dto.Name != null)
+                employee.Name = dto.Name;
+
+            if (dto.Salary.HasValue)
+                employee.UpdateSalary(dto.Salary.Value);
+
+
+            if (dto.Department != null)
+                employee.Department = dto.Department;
+
+            if (dto.Role != null)
+                employee.Role = dto.Role;
+
+            if (dto.Manager_Id.HasValue)
+                employee.manager_id = dto.Manager_Id.Value;
 
             await _repo.SaveAsync();
         }
 
         public async Task DeleteAsync(int id)
         {
+            var employee = await _repo.GetByIdAsync(id);
+            if (employee == null)
+                throw new KeyNotFoundException("Employee not found");
+
             await _repo.DeleteAsync(id);
+            _logger.LogInformation("Employee with Id {EmployeeId} deleted", id);
         }
